+++ backend/core/admin.py
@@ -2,7 +2,8 @@ from django.contrib import admin
 from django.utils import timezone
 from datetime import timedelta
 from django.contrib.admin import ModelAdmin
-from core.fileupload.models import File, Tag
+from core.fileupload.models.tag import Tag
+from core.fileupload.models.file import File
 from core.user.forms import AdminUserChangeForm, AdminUserCreationForm
 from core.user.models import User
 from django.contrib.auth.admin import UserAdmin as BaseUserAdmin
@@ -64,12 +65,12 @@ class FileAdmin(ModelAdmin):
     """
     model = File
     # Show attributes in list view
-    list_display = ('id', 'file', 'owner', 'uploaded_at')
+    list_display = ('id', 'new_version_of', 'local_file', 'owner', 'uploaded_at')
     # list_filter = ('is_superuser',)
     # Define how view should look like after clicking on an email
     fieldsets = [
         (None, {'fields': ['owner']}),
-        ('Information', {'fields': ['description', 'file', 'license']}),
+        ('Information', {'fields': ['description', 'local_file', 'license', 'tags', 'new_version_of']}),
         ('Important dates', {'fields': ['uploaded_at']})
     ]
     readonly_fields = ('uploaded_at',)
@@ -83,10 +84,10 @@ class TagAdmin(ModelAdmin):
     Class for defining the backend file admin panel and which data should be displayed.
     """
     model = Tag
-    list_display = ('id', 'label', 'creator', 'date_created')
+    list_display = ('id', 'label', 'is_public', 'creator', 'date_created')
     fieldsets = [
-        (None, {'fields': ['label']}),
-        ('Information', {'fields': ['id', 'creator', 'description']}),
+        (None, {'fields': ['id']}),
+        ('Information', {'fields': ['label', 'is_public', 'creator', 'description']}),
         ('Important dates', {'fields': ['date_created']})
     ]
     readonly_fields = ('id', 'date_created')
